Change confidence_level type in migration

+++ database/migrations/2022_04_08_094344_create_case_herpes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('case_herpes', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('disease_id');
-            $table->integer('confidence_level');
+            $table->double('confidence_level');
             $table->integer('age');
             $table->string('gender');
             $table->timestamps();
